+++ lib/grape/jsonapi.rb
@@ -1,9 +1,10 @@
-require 'grape/jsonapi/exceptions/filter_error'
+require 'grape/jsonapi/exceptions'
 require 'grape/jsonapi/document'
 require 'grape/jsonapi/exposer'
 require 'grape/jsonapi/entity'
 require 'grape/jsonapi/formatter'
 require 'grape/jsonapi/parameters'
+require 'grape/jsonapi/services'
 
 module Grape
   module Jsonapi