+++ lib/grape/jsonapi/exceptions.rb
@@ -1,0 +1,10 @@
+require 'grape/jsonapi/exceptions/filter_error'
+require 'grape/jsonapi/exceptions/unsupported_media_type_error'
+require 'grape/jsonapi/exceptions/not_acceptable_error'
+
+module Grape
+  module Jsonapi
+    module Exceptions
+    end
+  end
+end