+++ lib/grape/jsonapi/exceptions/not_acceptable_error.rb
@@ -1,0 +1,13 @@
+module Grape
+  module Jsonapi
+    module Exceptions
+      class NotAcceptableError < StandardError
+        MESSAGE = 'Accept header must be JSON API-compliant'.freeze
+
+        def initialize(msg = MESSAGE)
+          super
+        end
+      end
+    end
+  end
+end