+++ spec/grape/jsonapi/exceptions/not_acceptable_error_spec.rb
@@ -1,0 +1,8 @@
+describe Grape::Jsonapi::Exceptions::NotAcceptableError do
+  describe 'MESSAGE' do
+    it 'is the not acceptable error message' do
+      expect(Grape::Jsonapi::Exceptions::NotAcceptableError::MESSAGE)
+        .to eq('Accept header must be JSON API-compliant')
+    end
+  end
+end